requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0
PyYAML>=6.0.1
loguru>=0.7.2
pydantic>=2.5.2
typing-extensions>=4.8.0

# Development dependencies
pytest>=7.4.2
pytest-asyncio>=0.21.1
black>=23.9.1
isort>=5.12.0
mypy>=1.5.1
pylint>=3.0.2
